pyproject.toml
[project]
name = "mightex-slc-test"
version = "0.1.0"
description = "Python interface for Mightex SLC LED controller via RS232"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Logan Kaising"}
]

dependencies = [
    "pyserial>=3.5",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py39"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "hardware: tests that require a physical Mightex SLC controller on /dev/ttyUSB0",
]
addopts = "-m 'not hardware'"
